pyproject.toml
[project]
name = "quick-ocr-vlm-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "pyperclip>=1.9.0",
    "requests>=2.32.3",
]
